requirements.txt
elasticsearch==6.3.1
pika==0.12.0
python-dateutil==2.7.5
python-dotenv==0.10.0
six==1.12.0
urllib3==1.24.1
